% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{make_lorem_ipsum}
\alias{make_lorem_ipsum}
\title{Make a lorem ipsum}
\usage{
make_lorem_ipsum(words_count)
}
\arguments{
\item{words_count}{the number of words you would like to create}
}
\description{
Make a lorem ipsum
}
