% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{sample_letters}
\alias{sample_letters}
\title{Sample from letters}
\usage{
sample_letters(x)
}
\arguments{
\item{x}{the number of letters you would like to have}
}
\description{
Sample from letters
}
